Add mvp of signature generation from coordinates based on unimol_tools
This is the first step of moving this to an easy to install
package you can use the fine tuned unimol models we have
trained on our signatures without needign to install Uni-Core
and Uni-Mol which are large deep learning frameworks with
lots of bells and whistles for many types of models that include
logic for training, fine tuning and much much more.

This code is based on unimol_tools which is a set of tools
Uni-Mol have made to simplify some operations with the models.
I've torn out anything to do with the MOF models and many other
functions that are not needed to do inference with our
our fine tuned models (the signaturizers).

The tests pass currently, so this is the starting point
for converting unimol_tools into a useable and easy to
install package for signaturizers using the fine tuned unimol
models.
Files in /unicore/ are the needed deps copied from unicore.
The remainder is files from unimol_tools https://github.com/dptech-corp/Uni-Mol/tree/main/unimol_tools

The following work will be to remove any code not needed to
do inference with the fine tuned models from coordinates and smiles.
I expect to remove a lot of files and several of the dependencies
introduced here (pandas, scikit-learn, maybe more).

The dependencies needed to run this is in the pyproject.toml file,
excluding pytorch which needs to be installed separately for now.