Load model weights from github for all spaces
The weights are loaded from using torch model hub, so after the
first download the weights are cached locally.
The github repo used to host the weights is a temporary
location to test out how donwloading and caching the weights work.
It woudl be better to have them hosted in the same place as the
signaturizer3d code, but the assets upload on gitlab is limited
at 10MB so that won't work (the 25 space weighs total ~14GB), but
maybe creating a parallele repo on gitlab with the weights could work.
Github + release files have no size limit and it is the recommended
solution on the pytorch model hub docs (https://pytorch.org/docs/stable/hub.html).
Using releaese files means people cloning the repo won't need to download
the weights. This is important because the weights could quickly make
the git history of the repo very very large (14GB is already huge) if
we upload multiple versions of the weights with time.