Remove scipy dependency
Scipy is large, and we only use it for a single function.
We can do the same job with numpy, which is already a dependency.
The results differ slightly but I don't think the input
coordinates are this sensitive to change as the
input coordinates are not perfectly accurate anyway.