Only build new docker img when deps change (pyproject.toml or poetry.lock)
The docker image does not install the signaturizer3d package as the
code will change every commit even if the deps stay the same.
So any CI stages that use the signaturizer3d package have to run
poetry install to actually install the signaturizer3d package, but this
will be fast since all the deps are already installed.